requirements.txt
# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# Async HTTP client
aiohttp>=3.9.0

# Retry mechanism
tenacity>=8.2.0

# Telegram bot support
python-telegram-bot>=20.0

# Web search capabilities
duckduckgo-search>=4.1

# Logging and utilities
colorama>=0.4.6

# Future enhancements (optional)
# faiss-cpu>=1.7.0  # For vector search
# sentence-transformers>=2.2.0  # For embeddings

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
